pyproject.toml
[tool.poetry]
name = "spotify-idle"
version = "0.1.0"
description = ""
authors = ["darakudou <blackrx.robo.vaio@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
spotipy = "^2.16.1"
black = "^20.8b1"
autoflake = "^1.4"
flake8 = "^3.8.4"
pre-commit = "^2.7.1"
ipython = "^7.18.1"
retrying = "^1.3.3"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']
